requirements.txt
python-dotenv
pandas
groq
fastapi
uvicorn
openpyxl
streamlit
requests
numpy
scikit-learn
joblib

